refactor: API検証とモデル選択を別画面に分離
- ApiKeySetup.tsx をシンプル化（キー入力のみ）
- ModelValidation.tsx を新規作成（検証・モデル選択・オプション）
- 生体認証オプションを下部にまとめ
- モバイル対応のフルスクリーンレイアウト
- 戻るボタンを左上に配置
